// src/actions/types.ts
export enum actionTypes {
    authenticate,
    logout,
    authError,
    authLoading,

    showAlert,
    hideAlert,
    setNotifications,
    setFriends,
    setPendingInvitations,
    setOnlineUsers,
    setUserNotifcations,
    setGroupChatList,
    resetFriends,

    setChatType,
    setChosenChatDetails,
    setMessages,
    setGroupNotifications,
    addNewMessage,
    resetChat,
    setChosenGroupChatDetails,

    setTyping,
    setInitialTypingStatus,

    setLocalStream,
    setRemoteStream,
    setOtherUserId,
    setAudioOnly,
    setScreenSharingStream,
    setScreenSharing,
    setCallRequest,
    setCallStatus,
    resetVideoChatState,

    openRoom,
    setRoomDetails,
    setActiveRooms,
    setLocalStreamRoom,
    setRemoteStreams,
    setAudioOnlyRoom,
    setScreenSharingStreamRoom,
    setIsUserJoinedWithAudioOnly,
}

export interface CurrentUser {
    _id: string;
    email: string;
    token: string;
    username: string;
}
interface AuthSuccessAction {
    type: actionTypes.authenticate;
    payload: {
        _id: string;
        email: string;
        token: string;
        username: string;
    };
}

interface AuthErrorAction {
    type: actionTypes.authError;
    payload: string;
}
export interface OnlineNotifaction {
    department: string;
    totalNotifications: number;
}
interface LogoutAction {
    type: actionTypes.logout;
}
interface setNotifications {
    type: actionTypes.setNotifications;
    payload: Array<OnlineNotifaction>;
}
// ...
export interface GroupChatDetails {
    groupId: string;
    groupName: string;
    participants: Array<{
        _id: string;
        username: string;
        email: string;
    }>;
    admin: {
        _id: string;
        username: string;
        email: string;
    };
}
export interface UserGetNotification {
    userId: string,
    messages:Array<[]>;
}

interface SetPendingInvitationAction {
    type: actionTypes.setPendingInvitations;
    payload: Array<PendingInvitation>;
}

interface SetFriends {
    type: actionTypes.setFriends;
    payload: Array<Friend>;
}

interface SetOnlineUsers {
    type: actionTypes.setOnlineUsers;
    payload: Array<OnlineUser>;
}
interface setUserNotifcations {
    type: actionTypes.setUserNotifcations;
    payload: Array<UserGetNotification>;
}

export interface SetGroupChatList {
    type: actionTypes.setGroupChatList;
    payload: Array<GroupChatDetails>;
}

export interface SetChosenChatDetails {
    type: actionTypes.setChosenChatDetails;
    payload: {
        userId: string;
        username: string;
        department:string;
        section:string;
        designation:string;
        name: string;
    };
}

export interface SetChosenGroupChatDetails {
    type: actionTypes.setChosenGroupChatDetails;
    payload: GroupChatDetails;
}
export interface SetGroupNotifications {
    type: actionTypes.setGroupNotifications;
    payload: Array<{
        notifications: number;
        groupChatId: string;
    }>;
}

export interface ResetChat {
    type: actionTypes.resetChat;
}

export interface ResetFriends {
    type: actionTypes.resetFriends;
}

export interface Message {
    _id: string;
    content: string;
    author: {
        username: string;
        _id: string;
    };
    createdAt: string;
}
export interface total{
total:number
}
export interface Typing {
    typing: boolean;
    userId: string;
}

export interface SetMessages {
    type: actionTypes.setMessages;
    payload: Array<Message>;
}


export interface AddNewMessage {
    type: actionTypes.addNewMessage;
    payload: Message;
}

export interface SetTyping {
    type: actionTypes.setTyping;
    payload: {
        typing: boolean;
        userId: string;
    };
}

export interface SetInitialTypingStatus {
    type: actionTypes.setInitialTypingStatus;
    payload: Array<Typing>;
}
// ...
interface setOtherUserId {
    type: actionTypes.setOtherUserId;
    payload: {
        otherUserId: string | null;
    };
}

interface setScreenSharingStream {
    type: actionTypes.setScreenSharingStream;
    payload: {
        stream: MediaStream | null;
        isScreenSharing: boolean;
    };
}
// ...
export type ChatActions =
    | SetChosenChatDetails
    | SetChosenGroupChatDetails
    | SetMessages
    | AddNewMessage
    | SetTyping
    | SetGroupNotifications
    | SetInitialTypingStatus
    | ResetChat;